Split resource.py
