+++ setup.py
@@ -53,7 +53,9 @@ setup(
     author='Stripe',
     author_email='support@stripe.com',
     url='https://github.com/stripe/stripe-python',
-    packages=['stripe', 'stripe.test', 'stripe.test.resources'],
+    packages=['stripe', 'stripe.api_resources',
+              'stripe.api_resources.abstract',
+              'stripe.test', 'stripe.test.resources'],
     package_data={'stripe': ['data/ca-certificates.crt']},
     install_requires=install_requires,
     test_suite='stripe.test.all',
