+++ stripe/api_resources/abstract/api_resource.py
@@ -1,0 +1,44 @@
+import urllib
+
+from stripe import error, util
+from stripe.stripe_object import StripeObject
+
+
+class APIResource(StripeObject):
+
+    @classmethod
+    def retrieve(cls, id, api_key=None, **params):
+        instance = cls(id, api_key, **params)
+        instance.refresh()
+        return instance
+
+    def refresh(self):
+        self.refresh_from(self.request('get', self.instance_url()))
+        return self
+
+    @classmethod
+    def class_name(cls):
+        if cls == APIResource:
+            raise NotImplementedError(
+                'APIResource is an abstract class.  You should perform '
+                'actions on its subclasses (e.g. Charge, Customer)')
+        return str(urllib.quote_plus(cls.__name__.lower()))
+
+    @classmethod
+    def class_url(cls):
+        cls_name = cls.class_name()
+        return "/v1/%ss" % (cls_name,)
+
+    def instance_url(self):
+        id = self.get('id')
+
+        if not isinstance(id, basestring):
+            raise error.InvalidRequestError(
+                'Could not determine which URL to request: %s instance '
+                'has invalid ID: %r, %s. ID should be of type `str` (or'
+                ' `unicode`)' % (type(self).__name__, id, type(id)), 'id')
+
+        id = util.utf8(id)
+        base = self.class_url()
+        extn = urllib.quote_plus(id)
+        return "%s/%s" % (base, extn)
