+++ stripe/api_resources/abstract/deletable_api_resource.py
@@ -1,0 +1,8 @@
+from stripe.api_resources.abstract.api_resource import APIResource
+
+
+class DeletableAPIResource(APIResource):
+
+    def delete(self, **params):
+        self.refresh_from(self.request('delete', self.instance_url(), params))
+        return self
