+++ stripe/api_resources/abstract/singleton_api_resource.py
@@ -1,0 +1,16 @@
+from stripe.api_resources.abstract.api_resource import APIResource
+
+
+class SingletonAPIResource(APIResource):
+
+    @classmethod
+    def retrieve(cls, **params):
+        return super(SingletonAPIResource, cls).retrieve(None, **params)
+
+    @classmethod
+    def class_url(cls):
+        cls_name = cls.class_name()
+        return "/v1/%s" % (cls_name,)
+
+    def instance_url(self):
+        return self.class_url()
