+++ stripe/api_resources/apple_pay_domain.py
@@ -1,0 +1,12 @@
+from stripe.api_resources.abstract import CreateableAPIResource
+from stripe.api_resources.abstract import DeletableAPIResource
+from stripe.api_resources.abstract import ListableAPIResource
+
+
+class ApplePayDomain(CreateableAPIResource, ListableAPIResource,
+                     DeletableAPIResource):
+    OBJECT_NAME = 'apple_pay_domain'
+
+    @classmethod
+    def class_url(cls):
+        return '/v1/apple_pay/domains'
