+++ stripe/api_resources/balance.py
@@ -1,0 +1,5 @@
+from stripe.api_resources.abstract import SingletonAPIResource
+
+
+class Balance(SingletonAPIResource):
+    OBJECT_NAME = 'balance'
