+++ stripe/api_resources/balance_transaction.py
@@ -1,0 +1,9 @@
+from stripe.api_resources.abstract import ListableAPIResource
+
+
+class BalanceTransaction(ListableAPIResource):
+    OBJECT_NAME = 'balance_transaction'
+
+    @classmethod
+    def class_url(cls):
+        return '/v1/balance/history'
