+++ stripe/api_resources/bitcoin_transaction.py
@@ -1,0 +1,5 @@
+from stripe.stripe_object import StripeObject
+
+
+class BitcoinTransaction(StripeObject):
+    OBJECT_NAME = 'bitcoin_transaction'
