+++ stripe/api_resources/coupon.py
@@ -1,0 +1,9 @@
+from stripe.api_resources.abstract import CreateableAPIResource
+from stripe.api_resources.abstract import DeletableAPIResource
+from stripe.api_resources.abstract import UpdateableAPIResource
+from stripe.api_resources.abstract import ListableAPIResource
+
+
+class Coupon(CreateableAPIResource, UpdateableAPIResource,
+             DeletableAPIResource, ListableAPIResource):
+    OBJECT_NAME = 'coupon'
