+++ stripe/api_resources/event.py
@@ -1,0 +1,5 @@
+from stripe.api_resources import abstract
+
+
+class Event(abstract.ListableAPIResource):
+    OBJECT_NAME = 'event'
