+++ stripe/api_resources/invoice_item.py
@@ -1,0 +1,9 @@
+from stripe.api_resources.abstract import CreateableAPIResource
+from stripe.api_resources.abstract import DeletableAPIResource
+from stripe.api_resources.abstract import UpdateableAPIResource
+from stripe.api_resources.abstract import ListableAPIResource
+
+
+class InvoiceItem(CreateableAPIResource, UpdateableAPIResource,
+                  ListableAPIResource, DeletableAPIResource):
+    OBJECT_NAME = 'invoice_item'
