+++ stripe/api_resources/login_link.py
@@ -1,0 +1,5 @@
+from stripe.stripe_object import StripeObject
+
+
+class LoginLink(StripeObject):
+    OBJECT_NAME = 'login_link'
