+++ stripe/api_resources/order_return.py
@@ -1,0 +1,9 @@
+from stripe.api_resources.abstract import ListableAPIResource
+
+
+class OrderReturn(ListableAPIResource):
+    OBJECT_NAME = 'order_return'
+
+    @classmethod
+    def class_url(cls):
+        return '/v1/order_returns'
