+++ stripe/api_resources/product.py
@@ -1,0 +1,9 @@
+from stripe.api_resources.abstract import CreateableAPIResource
+from stripe.api_resources.abstract import DeletableAPIResource
+from stripe.api_resources.abstract import UpdateableAPIResource
+from stripe.api_resources.abstract import ListableAPIResource
+
+
+class Product(CreateableAPIResource, UpdateableAPIResource,
+              ListableAPIResource, DeletableAPIResource):
+    OBJECT_NAME = 'product'
