+++ stripe/api_resources/refund.py
@@ -1,0 +1,8 @@
+from stripe.api_resources.abstract import CreateableAPIResource
+from stripe.api_resources.abstract import UpdateableAPIResource
+from stripe.api_resources.abstract import ListableAPIResource
+
+
+class Refund(CreateableAPIResource, ListableAPIResource,
+             UpdateableAPIResource):
+    OBJECT_NAME = 'refund'
