+++ stripe/api_resources/three_d_secure.py
@@ -1,0 +1,9 @@
+from stripe.api_resources.abstract import CreateableAPIResource
+
+
+class ThreeDSecure(CreateableAPIResource):
+    OBJECT_NAME = 'three_d_secure'
+
+    @classmethod
+    def class_url(cls):
+        return '/v1/3d_secure'
