+++ stripe/api_resources/token.py
@@ -1,0 +1,5 @@
+from stripe.api_resources.abstract import CreateableAPIResource
+
+
+class Token(CreateableAPIResource):
+    OBJECT_NAME = 'token'
